# docs/apis/prl-hub/openapi.v0.1.0.yaml
openapi: 3.0.3
info:
  title: PRL Hub API
  version: 0.1.0
  description: |
    Borrador versionado de la API REST del Hub PRL. Expone endpoints sincrónicos para
    altas y actualizaciones de pacientes, captura de cuestionarios clínicos, planificación
    de reconocimientos y orquestación de procesos de sincronización offline.
  contact:
    name: Equipo de integración PRL Hub
    email: integraciones@prevengos.test
  license:
    name: CC BY-NC-SA 4.0
servers:
  - url: https://hub.prl.local/api
    description: Entorno interno de referencia
  - url: https://hub-prl-pre.int.prevengos.test/api
    description: Preproducción integrada con Prevengos
  - url: https://hub-prl.prevengos.com/api
    description: Producción
...
paths:
...
  /sincronizacion/pull:
    get:
      tags: [Sincronizacion]
      summary: Obtiene cambios incrementales a partir de un token
      parameters:
        - in: query
          name: syncToken
          description: Último token consumido por el cliente. Si no se indica se parte desde el origen.
          required: false
          schema:
            type: integer
            format: int64
        - in: query
          name: limit
          description: Máximo de elementos por página (hasta 500)
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 500
      responses:
        '200':
          description: Listado incremental de pacientes, cuestionarios y eventos con el siguiente token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncPullResponse'
      security:
        - ApiKeyAuth: []
  /sincronizacion/push:
    post:
      tags: [Sincronizacion]
      summary: Entrega lote de altas/actualizaciones desde un cliente offline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncPushRequest'
      responses:
        '200':
          description: Lote procesado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncPushResponse'
        '409':
          description: Evento duplicado o conflicto de versión
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - ApiKeyAuth: []
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              issue:
                type: string
...
    SyncPullResponse:
      type: object
      required: [pacientes, cuestionarios, events, nextSyncToken]
      properties:
        pacientes:
          type: array
          items:
            $ref: '#/components/schemas/PacienteSync'
        cuestionarios:
          type: array
          items:
            $ref: '#/components/schemas/CuestionarioSync'
        events:
          type: array
          items:
            $ref: '#/components/schemas/SyncEvent'
        nextSyncToken:
          type: integer
          format: int64
    SyncPushRequest:
      type: object
      required: [source, pacientes, cuestionarios]
      properties:
        source:
          type: string
        correlationId:
          type: string
          format: uuid
        pacientes:
          type: array
          items:
            $ref: '#/components/schemas/PacienteSync'
        cuestionarios:
          type: array
          items:
            $ref: '#/components/schemas/CuestionarioSync'
    SyncPushResponse:
      type: object
      required: [processedPacientes, processedCuestionarios, lastSyncToken]
      properties:
        processedPacientes:
          type: integer
          format: int64
        processedCuestionarios:
          type: integer
          format: int64
        lastSyncToken:
          type: integer
          format: int64
        createdOrUpdatedIds:
          type: array
          items:
            type: string
            format: uuid
    PacienteSync:
      type: object
      required: [pacienteId, nombre, apellidos]
      properties:
        pacienteId:
          type: string
          format: uuid
        nif:
          type: string
        nombre:
          type: string
        apellidos:
          type: string
        fechaNacimiento:
          type: string
          format: date
        sexo:
          type: string
          maxLength: 1
        telefono:
          type: string
        email:
          type: string
          format: email
        empresaId:
          type: string
          format: uuid
        centroId:
          type: string
          format: uuid
        externoRef:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastModified:
          type: string
          format: date-time
        syncToken:
          type: integer
          format: int64
    CuestionarioSync:
      type: object
      required: [cuestionarioId, pacienteId, plantillaCodigo, estado]
      properties:
        cuestionarioId:
          type: string
          format: uuid
        pacienteId:
          type: string
          format: uuid
        plantillaCodigo:
          type: string
        estado:
          type: string
        respuestas:
          type: string
          description: Cadena JSON con las respuestas capturadas.
        firmas:
          type: string
          description: Cadena JSON con las firmas asociadas.
        adjuntos:
          type: string
          description: Cadena JSON con referencias a adjuntos.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastModified:
          type: string
          format: date-time
        syncToken:
          type: integer
          format: int64
    SyncEvent:
      type: object
      required: [syncToken, eventId, eventType, version, occurredAt, source, payload, metadata]
      properties:
        syncToken:
          type: integer
          format: int64
        eventId:
          type: string
          format: uuid
        eventType:
          type: string
        version:
          type: integer
        occurredAt:
          type: string
          format: date-time
        source:
          type: string
        correlationId:
          type: string
          format: uuid
          nullable: true
        causationId:
          type: string
          format: uuid
          nullable: true
        payload:
          type: string
          description: Payload serializado del evento.
        metadata:
          type: string
          description: Metadata adicional serializada.
...